PP-4002 Move GoCardlessCreditor* subtypes to dedicated package
with @danworth
